% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{\%>>=\%}
\alias{\%>>=\%}
\alias{bind}
\title{The 'bind' operator}
\usage{
lhs \%>>=\% rhs
}
\arguments{
\item{lhs}{The left-hand side of the operator.}

\item{rhs}{Although not required, it is intended that the right-hand side
contain a function or block, evaluated with \code{.} bound to the value of
\code{lhs}.}
}
\description{
\code{\%>>=\%} (pronounced "bind") is a generic operator intended to match
the semantics of \code{>>=} in other functional programming languages, such as F#
or Haskell. It is intended that the right-hand side will contain a function
or block, similar to the pipe operator \code{\link[magrittr]{\%>\%}}.
}
