% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{\%>>=\%}
\alias{\%>>=\%}
\title{The 'bind' operator}
\usage{
lhs \%>>=\% rhs
}
\description{
\code{\%>>=\%} (pronounced "bind") is a generic operator intended to match
the semantics of \code{>>=} in other functional programming languages, such
as F# or Haskell.
}

