% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{as_result}
\alias{as_result}
\alias{as_result.default}
\alias{as_result.error}
\alias{as_result.result}
\alias{is_result}
\alias{result_is_ok}
\alias{result_is_error}
\alias{print.result}
\alias{unwrap_result}
\title{The Result object}
\usage{
as_result(x, ...)

\method{as_result}{default}(x, ...)

\method{as_result}{error}(x, ...)

\method{as_result}{result}(x, ...)

is_result(x)

result_is_ok(x)

result_is_error(x)

\method{print}{result}(x, ...)

unwrap_result(x)
}
\arguments{
\item{x}{An object.}

\item{...}{Further arguments passed on to other methods (unused by the
default methods).}
}
\description{
\code{is_result} and \code{as_result} test whether an object is a result and
coerce to a result object, respectively.
}
