% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{result}
\alias{result}
\alias{as_result}
\alias{as_result.default}
\alias{as_result.error}
\alias{as_result.result}
\alias{is_result}
\alias{print.result}
\alias{result_is_ok}
\alias{result_is_ok.result}
\alias{result_is_ok.list}
\alias{result_is_error}
\alias{result_is_error.result}
\alias{result_is_error.list}
\alias{unwrap}
\alias{unwrap.result}
\alias{unwrap.list}
\title{The Result object}
\usage{
as_result(x, ...)

\method{as_result}{default}(x, ...)

\method{as_result}{error}(x, ...)

\method{as_result}{result}(x, ...)

is_result(x)

\method{print}{result}(x, ...)

result_is_ok(x)

\method{result_is_ok}{result}(x)

\method{result_is_ok}{list}(x)

result_is_error(x)

\method{result_is_error}{result}(x)

\method{result_is_error}{list}(x)

unwrap(x)

\method{unwrap}{result}(x)

\method{unwrap}{list}(x)
}
\arguments{
\item{x}{An object.}

\item{...}{Further arguments passed on to other methods (unused by the
default methods).}
}
\description{
\code{is_result} and \code{as_result} test whether an object is a result and
coerce to a result object, respectively.
}
